Update csrf() helper docs
The csrf() helper returns the CSRF token if it is called the first time. If the developer forgot to add the CSRF form field, the helper will return the token in the if condition, which evaluates to true. That's obviously not what we want so we should do a strict check here.